// Src/ShopperHolic.WebUI/ShopperHolic/src/app/models/accounts/customers/customer.ts
export class Customer
{
    customerID: number = 0; 
    customerTypeID: number = 0; 
    customerTypeText: string = ""; 
    defaultAddressID: number = 0; 
    defaultAddressText: string = ""; 
    customerCode: string = ""; 
    customerName: string = ""; 
    customerContactNumber: string = ""; 
    customerEmailAddress: string = "";
}


 